Seat UI 2.0

+++ Service/EmployeeService.cs
@@ -1,5 +1,4 @@
-﻿using Microsoft.EntityFrameworkCore;
-using MovieTheater.Models;
+﻿using MovieTheater.Models;
 using MovieTheater.Repository;
 using MovieTheater.ViewModels;
 
@@ -9,13 +8,11 @@ namespace MovieTheater.Service
     {
         private readonly IEmployeeRepository _repository;
         private readonly IAccountService _accountService;
-        private readonly IAccountRepository _accountRepo;
 
-        public EmployeeService(IEmployeeRepository repository, IAccountService accountService, IAccountRepository accountRepo)
+        public EmployeeService(IEmployeeRepository repository, IAccountService accountService)
         {
             _repository = repository;
             _accountService = accountService;
-            _accountRepo = accountRepo;
         }
 
         public IEnumerable<Employee> GetAll()
@@ -36,7 +33,7 @@ namespace MovieTheater.Service
             return result;
         }
 
-        
+
         public bool Update(string id, RegisterViewModel model)
         {
             var employee = _repository.GetById(id);
@@ -47,6 +44,8 @@ namespace MovieTheater.Service
 
         public bool Delete(string employeeId)
         {
+            var employee = _repository.GetById(employeeId);
+            if (employee == null) return false;
             _repository.Delete(employeeId);
             _repository.Save();
             return true;
